requirements.txt
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
torch
stable-baselines3
gymnasium
PyYAML
SQLAlchemy
alembic
psycopg[binary]
streamlit
plotly
matplotlib
spacy
keybert
reportlab
rich
click
python-dotenv
fastapi
uvicorn
